+++ CHANGES.md
@@ -2,12 +2,28 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## photos-app:0.1.1
+## 2021-05-21
+
+### photos-app:0.2.0
+* Change HTML tag depending on photo of video media
+* Send event to update device "mediaType" attribute
+
+### photos-device:0.0.2
+* New attribute: mediaType
+
+## 2021-05-17
+
+### photos-app:0.1.1
 * Bugfix: nightly refresh runs every second for the hour of 11pm
 
-## photos-app:0.1.0
+### photos-app:0.1.0
 * Set refresh interval via App preferences
 * Preferences toggle to refresh list of photo IDs in album nightly (at 11pm)
 
-## photos-app:0.0.2
+## 2021-05-15
+
+### photos-app:0.0.2
 * Initial implementation -- authorization, API queries
+
+### photos-device:0.0.1
+* Initial implementation -- image attribute, refresh command